requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
requests==2.31.0
sniffio==1.3.0
urllib3==2.2.1
neo4j==5.26.0
retry==0.9.2
python-dotenv==1.0.0
requests
streamlit
python-dotenv
asyncio
fastapi
langchain
langchain-core
langchain-openai
langchainhub
neo4j
numpy
openai
opentelemetry-api
pydantic
uvicorn
langchain-neo4j
python-dotenv
langchain_community 



